stop loading the entire library when disabled
