+++ packages/dd-trace/src/index.js
@@ -1,0 +1,7 @@
+'use strict'
+
+const { isFalse } = require('./util')
+
+module.exports = isFalse(process.env.DD_TRACE_ENABLED)
+  ? require('./noop/proxy')
+  : require('./proxy')
