+++ packages/dd-trace/src/proxy.js
@@ -1,27 +1,25 @@
 'use strict'
 
-const NoopTracer = require('./noop/tracer')
+const NoopProxy = require('./noop/proxy')
 const DatadogTracer = require('./tracer')
 const Config = require('./config')
 const metrics = require('./metrics')
 const log = require('./log')
-const { isFalse } = require('./util')
 const { setStartupLogPluginManager } = require('./startup-log')
 const telemetry = require('./telemetry')
 const PluginManager = require('./plugin_manager')
 const { sendGitMetadata } = require('./ci-visibility/exporters/git/git_metadata')
 
-const noop = new NoopTracer()
-
-class Tracer {
+class Tracer extends NoopProxy {
   constructor () {
+    super()
+
     this._initialized = false
-    this._tracer = noop
     this._pluginManager = new PluginManager(this)
   }
 
   init (options) {
-    if (isFalse(process.env.DD_TRACE_ENABLED) || this._initialized) return this
+    if (this._initialized) return this
 
     this._initialized = true
 
@@ -76,61 +74,6 @@ class Tracer {
   use () {
     this._pluginManager.configurePlugin(...arguments)
     return this
-  }
-
-  trace (name, options, fn) {
-    if (!fn) {
-      fn = options
-      options = {}
-    }
-
-    if (typeof fn !== 'function') return
-
-    options = options || {}
-
-    return this._tracer.trace(name, options, fn)
-  }
-
-  wrap (name, options, fn) {
-    if (!fn) {
-      fn = options
-      options = {}
-    }
-
-    if (typeof fn !== 'function') return fn
-
-    options = options || {}
-
-    return this._tracer.wrap(name, options, fn)
-  }
-
-  setUrl () {
-    this._tracer.setUrl.apply(this._tracer, arguments)
-    return this
-  }
-
-  startSpan () {
-    return this._tracer.startSpan.apply(this._tracer, arguments)
-  }
-
-  inject () {
-    return this._tracer.inject.apply(this._tracer, arguments)
-  }
-
-  extract () {
-    return this._tracer.extract.apply(this._tracer, arguments)
-  }
-
-  scope () {
-    return this._tracer.scope.apply(this._tracer, arguments)
-  }
-
-  getRumData () {
-    return this._tracer.getRumData.apply(this._tracer, arguments)
-  }
-
-  setUser () {
-    return this._tracer.setUser.apply(this.tracer, arguments)
   }
 }
